+++ anyblok/config.py
@@ -17,6 +17,7 @@ import os
 import json
 import yaml
 import urllib
+import warnings
 from appdirs import AppDirs
 from os.path import join, isfile
 from sqlalchemy.engine.url import URL, make_url
@@ -198,16 +199,22 @@ def nargs_type(key, nargs, cast):
 
 class AnyBlokActionsContainer:
 
-    def add_argument(self, *args, **kwargs):
+    def add_argument(self, *args, deprecated=None, **kwargs):
         """Overload the method to add the entry in the configuration dict
 
         :param args:
         :param kwargs:
+        :param deprecated: Deprecated message to display
         :return:
         """
         default = kwargs.pop('default', None)
         nargs = kwargs.get('nargs', None)
         type = kwargs.get('type')
+
+        if deprecated is not None:
+            kwargs['help'] = '[DEPRECATED] {} : {}'.format(
+                kwargs.get('help', ''), deprecated)
+
         arg = super(AnyBlokActionsContainer, self).add_argument(
             *args, **kwargs)
         if type is None:
@@ -222,7 +229,8 @@ class AnyBlokActionsContainer:
         if nargs:
             type = nargs_type(dest, nargs, type)
 
-        Configuration.add_argument(dest, default, type=type)
+        Configuration.add_argument(dest, default, type=type,
+                                   deprecated=deprecated)
         return arg
 
     def set_defaults(self, **kwargs):
@@ -286,8 +294,9 @@ def AnyBlokPlugin(import_definition):
 
 class ConfigOption:
 
-    def __init__(self, value, type):
+    def __init__(self, value, type, deprecated=None):
         self.type = type
+        self.deprecated = deprecated
         self.set(value)
 
     def get(self):
@@ -295,6 +304,9 @@ class ConfigOption:
 
         :return:
         """
+        if self.deprecated is not None:
+            warnings.warn(self.deprecated, DeprecationWarning, stacklevel=2)
+
         return self.value
 
     def set(self, value):
@@ -302,6 +314,9 @@ class ConfigOption:
 
         :param value:
         """
+        if self.deprecated is not None:
+            warnings.warn(self.deprecated, DeprecationWarning, stacklevel=2)
+
         self.value = cast_value(self.type, value)
 
 
@@ -517,14 +532,16 @@ class Configuration:
             cls.set(k, v)
 
     @classmethod
-    def add_argument(cls, key, value, type=str):
+    def add_argument(cls, key, value, type=str, deprecated=None):
         """Add a configuration option
 
         :param key:
         :param value:
         :param type:
+        :param deprecated: deprecated message to warn in get/set methods
         """
-        cls.configuration[key] = ConfigOption(value, type)
+        cls.configuration[key] = ConfigOption(
+            value, type, deprecated=deprecated)
 
     @classmethod
     def remove_label(cls, group):
@@ -778,15 +795,24 @@ def add_plugins(group):
     """
     group.add_argument('--registry-cls', dest='Registry', type=AnyBlokPlugin,
                        default='anyblok.registry:Registry',
-                       help="Registry class to use")
+                       help="Registry class to use",
+                       deprecated=("This plugins will be removed in version "
+                                   "2.0, because this behaviours is "
+                                   "dangerous"))
     group.add_argument('--migration-cls', dest='Migration',
                        type=AnyBlokPlugin,
                        default='anyblok.migration:Migration',
-                       help="Migration class to use")
+                       help="Migration class to use",
+                       deprecated=("This plugins will be removed in version "
+                                   "2.0, use the entry point "
+                                   "'anyblok.migration_type.plugins'"))
     group.add_argument('--get-url-fnct', dest='get_url',
                        type=AnyBlokPlugin,
                        default='anyblok.config:get_url',
-                       help="get_url function to use")
+                       help="get_url function to use",
+                       deprecated=("This plugins will be removed in version "
+                                   "2.0, because this behaviours is "
+                                   "dangerous"))
 
 
 @Configuration.add('config', must_be_loaded_by_unittest=True)